backend/requirements-dev.txt
# Development Requirements
-r requirements.txt

# Development Tools
django-debug-toolbar==4.2.0
django-silk==5.0.4
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
httpx==0.25.2

# Data Management
django-import-export==3.3.7

# Compression
django-compressor==4.4